feat: Basic comic navigation implemented
There's a basic page layout consisting of the typical header, sidebar
and main content area. The sidebar is filled from a folder on the
local computer that the user must select and give access to.
The user can then select Comic pages which will be displayed in the
main content area.